Sync: most PostgresReplicationListener to sync package
