+++ packages/sync/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@rejot/sync",
+  "version": "0.0.0",
+  "scripts": {
+    "check": "tsc --noEmit"
+  },
+  "exports": {
+    "./postgres": "./src/postgres/postgres-replication-listener.ts"
+  },
+  "private": true,
+  "type": "module",
+  "dependencies": {
+    "pg": "^8.13.3",
+    "pg-logical-replication": "^2.0.7"
+  }
+}
